+++ Game/src/savegamestorage.cpp
@@ -398,8 +398,8 @@ std::filesystem::path SaveGameStorage::GetSaveGameFileName( const std::string& c
 	// These characters aren't supported by the operative system. Having them as part of the filename would
 	// cause the save game to fail, so we replace them with a '_'
 	const std::string invalidCharacters( "\\/*?\"<>|" );
-	const int filenameLength = filename.length();
-	const int invalidCharactersLength = invalidCharacters.length();
+	const int filenameLength = static_cast<int>(filename.length());
+	const int invalidCharactersLength = static_cast<int>(invalidCharacters.length());
 	for ( int i = 0; i < filenameLength; ++i )
 	{
 		for ( int j = 0; j < invalidCharactersLength; ++j )
